cleaning .cs files
